+++ src/main/java/kitchenpos/service/MenuGroupService.java
@@ -1,11 +1,12 @@
-package kitchenpos.application;
+package kitchenpos.service;
 
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.domain.MenuGroup;
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
+import kitchenpos.dao.MenuGroupDao;
+import kitchenpos.domain.MenuGroup;
 
 @Service
 public class MenuGroupService {
