+++ src/main/java/kitchenpos/service/ProductService.java
@@ -1,13 +1,14 @@
-package kitchenpos.application;
-
-import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.Product;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
+package kitchenpos.service;
 
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
+
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import kitchenpos.dao.ProductDao;
+import kitchenpos.domain.Product;
 
 @Service
 public class ProductService {
