+++ src/main/java/kitchenpos/web/MenuGroupRestController.java
@@ -1,15 +1,16 @@
-package kitchenpos.ui;
+package kitchenpos.web;
 
-import kitchenpos.application.MenuGroupService;
-import kitchenpos.domain.MenuGroup;
+import java.net.URI;
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
-import java.util.List;
+import kitchenpos.domain.MenuGroup;
+import kitchenpos.service.MenuGroupService;
 
 @RestController
 public class MenuGroupRestController {
