+++ src/main/java/kitchenpos/web/MenuRestController.java
@@ -1,15 +1,16 @@
-package kitchenpos.ui;
+package kitchenpos.web;
 
-import kitchenpos.application.MenuService;
-import kitchenpos.domain.Menu;
+import java.net.URI;
+import java.util.List;
+
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.net.URI;
-import java.util.List;
+import kitchenpos.domain.Menu;
+import kitchenpos.service.MenuService;
 
 @RestController
 public class MenuRestController {
