+++ src/main/java/kitchenpos/web/TableGroupRestController.java
@@ -1,11 +1,16 @@
-package kitchenpos.ui;
-
-import kitchenpos.application.TableGroupService;
-import kitchenpos.domain.TableGroup;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+package kitchenpos.web;
 
 import java.net.URI;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+import kitchenpos.domain.TableGroup;
+import kitchenpos.service.TableGroupService;
 
 @RestController
 public class TableGroupRestController {
